Add token condition

+++ api/1.0/category/delete.php
@@ -8,26 +8,33 @@ header('Access-Control-Allow-Headers: Access-Control-Allow-Headers, Content-type
 include_once '../config/Database.php';
 include_once '../models/Category.php';
 
-// Instantiate DB + connect
-$database = new Database();
-$db = $database->connect();
+// Get token and continue
+if ($_SERVER['HTTP_TOKEN'] != '' && $_SERVER['HTTP_TOKEN'] == '@@admintoken') {
+    // Instantiate DB + connect
+    $database = new Database();
+    $db = $database->connect();
 
-// Instantiate Category object
-$category = new Category($db);
+    // Instantiate Category object
+    $category = new Category($db);
 
-// Get raw Categoryed data
-$data = json_decode(file_get_contents("php://input"));
+    // Get raw Categoryed data
+    $data = json_decode(file_get_contents("php://input"));
 
-// Set id to update
-$category->id = $data->id;
+    // Set id to update
+    $category->id = $data->id;
 
-// Delete Category
-if ($category->delete()) {
-    echo json_encode(
-        array('message' => 'Category deleted')
-    );
-} else {
+    // Delete Category
+    if ($category->delete()) {
+        echo json_encode(
+            array('message' => 'Category deleted')
+        );
+    } else {
         echo json_encode(
             array('message' => 'Category not deleted')
         );
+    }
+} else {
+    echo json_encode(
+        array('error' => 'Wrong token!')
+    );
 }
